requirements.txt
streamlit==1.45.1
langchain==0.3.25
langchain-community==0.3.24
pypdf==5.5.0
openpyxl==3.1.5
pytesseract
pillow
langchain-huggingface
langchain-chroma>=0.1.2
python-dotenv
langchain-cerebras

